first commit I threw together a raytracer that can do among other things a raytrace through a simplified objective. Then there is a foreign function interface for fftw that I use to do 3d ffts. I also have fftshift3, do-box and storage of stacks into a directory.
There is a 3d bresenham algorithm to draw lines through a volume. There is code to calculate the amplitude point spread function of a high NA objective and also some code that modifies the PSF as the back focal plane is blocked.

There is simulated annealing combined with downhill simplex - a very simple optimization method.

There are some problems:
- Right now the angular-psf function doesn't block the back focal plane with the coordinates I want.
- There is a lot of overlap in the vector math. In run.lisp I introduce vec-i a 3 vector with fixnum components and at least one other file I have vec (with double-float components) and the corresponding functions. One day I should factor this out.